requirements.txt
fastapi[standard]
sqlalchemy
uvicorn
pytest
alembic
psycopg
psycopg-binary
python-dotenv
httpx